several logic updates
