+++ webui/react/src/components/kit/LineChart.tsx
@@ -13,7 +13,12 @@ import { closestPointPlugin } from 'components/kit/internal/UPlot/UPlotChart/clo
 import { tooltipsPlugin } from 'components/kit/internal/UPlot/UPlotChart/tooltipsPlugin';
 import useResize from 'components/kit/internal/useResize';
 import { XAxisDomain, XAxisFilter } from 'components/kit/LineChart/XAxisFilter';
-import { Loadable } from 'utils/loadable';
+import MetricBadgeTag from 'components/MetricBadgeTag';
+import { MapOfIdsToColors } from 'hooks/useGlasbey';
+import { TrialMetricData } from 'pages/TrialDetails/useTrialMetrics';
+import { ExperimentWithTrial, TrialItem } from 'types';
+import { isEqual } from 'utils/data';
+import { Loadable, Loaded, NotLoaded } from 'utils/loadable';
 
 import css from './LineChart.module.scss';
 
@@ -308,6 +313,71 @@ const VirtualChartRenderer: React.FC<
       </div>
     </div>
   );
+};
+
+export const calculateChartProps = (
+  metricData: TrialMetricData,
+  experiments: ExperimentWithTrial[],
+  trials: TrialItem[],
+  xAxis: XAxisDomain,
+  colorMap: MapOfIdsToColors,
+): Loadable<ChartProps> => {
+  const { metricHasData, metrics, data, isLoaded, selectedMetrics } = metricData;
+  const chartedMetrics: Record<string, boolean> = {};
+  const out: ChartsProps = [];
+  const expNameById: Record<number, string> = {};
+  experiments.forEach((e) => {
+    expNameById[e.experiment.id] = e.experiment.name;
+  });
+  metrics.forEach((metric) => {
+    const series: Serie[] = [];
+    const key = `${metric.type}|${metric.name}`;
+    trials.forEach((t) => {
+      const m = data[t?.id || 0];
+      m?.[key] &&
+        t &&
+        series.push({
+          ...m[key],
+          color: colorMap[t.experimentId],
+          metricType: undefined,
+          name: expNameById[t.experimentId]
+            ? `${expNameById[t.experimentId]} (${t.experimentId})`
+            : String(t.experimentId),
+        });
+      chartedMetrics[key] ||= series.length > 0;
+    });
+    out.push({
+      series: Loaded(series),
+      title: <MetricBadgeTag metric={metric} />,
+      xAxis,
+      xLabel: String(xAxis),
+    });
+  });
+
+  // In order to show the spinner for each chart in the ChartGrid until
+  // metrics are visible, we must determine whether the metrics have been
+  // loaded and whether the chart props have been updated.
+  // If any metric has data but no chartProps contain data for the metric,
+  // then the charts have not been updated and we need to continue to show the
+  // spinner.
+  const chartDataIsLoaded = metrics.every((metric) => {
+    const metricKey = `${metric.type}|${metric.name}`;
+    return metricHasData?.[metricKey] ? !!chartedMetrics?.[metricKey] : true;
+  });
+  if (!isLoaded) {
+    // When trial metrics hasn't loaded metric names or individual trial metrics.
+    return NotLoaded;
+  } else if (!chartDataIsLoaded || !isEqual(selectedMetrics, metrics)) {
+    // In some cases the selectedMetrics returned may not be up to date
+    // with the metrics selected by the user. In this case we want to
+    // show a loading state until the metrics match.
+
+    // returns the chartProps with a NotLoaded series which enables
+    // the ChartGrid to show a spinner for the loading charts.
+    return Loaded(out.map((chartProps) => ({ ...chartProps, series: NotLoaded })));
+  } else {
+    return Loaded(out);
+  }
 };
 
 export const ChartGrid: React.FC<GroupProps> = React.memo(
